requirements.txt
lxml==3.8.0
requests==2.18.3
scrapy-proxies==0.3
